+++ src/main/java/com/boatupgrades/BoatUpgradesPlugin.java
@@ -2,13 +2,13 @@ package com.boatupgrades;
 
 import com.google.inject.Provides;
 import javax.inject.Inject;
+
 import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.ChatMessageType;
 import net.runelite.api.Client;
 import net.runelite.api.GameState;
 import net.runelite.api.events.CommandExecuted;
 import net.runelite.api.events.GameStateChanged;
-import net.runelite.api.events.GameTick;
 import net.runelite.api.gameval.VarbitID;
 import net.runelite.client.callback.ClientThread;
 import net.runelite.client.config.ConfigManager;
@@ -39,7 +39,6 @@ public class BoatUpgradesPlugin extends Plugin
 	private Client client;
 	@Inject
 	private BoatUpgradesConfig config;
-	@Inject
 	private BoatUpgradesOverlay boatUpgradesOverlay;
 	@Inject
 	private OverlayManager overlayManager;
@@ -53,7 +52,6 @@ public class BoatUpgradesPlugin extends Plugin
 	private FacilityService facilityService;
 	@Inject
 	private ClientToolbar clientToolbar;
-	@Inject
 	private BoatUpgradesPanel panel;
 	private NavigationButton navButton;
 
@@ -62,6 +60,9 @@ public class BoatUpgradesPlugin extends Plugin
 	@Override
 	protected void startUp() throws Exception
 	{
+		// Inject after startup so that the BoatUpgradesPanel dependency is created
+		// after default SwingUI styling is applied
+		if (boatUpgradesOverlay == null) boatUpgradesOverlay = injector.getInstance(BoatUpgradesOverlay.class);
 		overlayManager.add(boatUpgradesOverlay);
 		facilityService.start();
 		log.info("Boat Upgrades started");
@@ -74,7 +75,7 @@ public class BoatUpgradesPlugin extends Plugin
 				.tooltip("Boat Upgrades")
 				.icon(icon)
 				.panel(panel)
-				//.priority(5)
+				.priority(config.panelPosition())
 				.build();
 
 		clientToolbar.addNavigation(navButton);
@@ -150,6 +151,29 @@ public class BoatUpgradesPlugin extends Plugin
 		if ("persistMinutes".equals(event.getKey()))
 		{
 			boatUpgradesOverlay.refreshCacheExpiry();
+		}
+		if ("filterSchematicRequirement".equals(event.getKey()) || "filterConstructionRequirement".equals(event.getKey()))
+		{
+			clientThread.invokeLater(() ->
+			{
+				panel.onAvailableUpgradesChanged();
+			});
+		}
+		if (event.getKey().equals("panelPosition"))
+		{
+			panel = injector.getInstance(BoatUpgradesPanel.class);
+			final BufferedImage icon = ImageUtil.loadImageResource(getClass(), "icon.png");
+
+				clientToolbar.removeNavigation(navButton);
+
+				navButton = NavigationButton.builder()
+						.tooltip("Boat Upgrades")
+						.icon(icon)
+						.panel(panel)
+						.priority(config.panelPosition())
+						.build();
+
+				clientToolbar.addNavigation(navButton);
 		}
 	}