+++ src/main/resources/changelog.md
@@ -1,3 +1,9 @@
+## 1.2.0
+- Added side panel for displaying available upgrades
+- Upgrade names and materials in the side panel can be clicked and will link to their wiki page
+- Added setting for specifying position of the side panel
+- Added setting for showing text overlay only when in the shipyard
+
 ## 1.1.0
 - Added boat facilities - enable them in the plugin settings
 - Added setting for enabling/disabling the text overlay